Refactor backend and frontend components.
