+++ backend/server.js
@@ -2,8 +2,11 @@ const express = require('express');
 const http = require('http');
 const socketIo = require('socket.io');
 const cors = require('cors');
-const { v4: uuidv4 } = require('uuid');
 const path = require('path');
+
+// Import services and handlers
+const meetingsRoutes = require('./routes/meetings');
+const socketHandlers = require('./handlers/socketHandlers');
 
 const app = express();
 const server = http.createServer(app);
@@ -18,7 +21,7 @@ const corsOptions = {
 app.use(cors(corsOptions));
 app.use(express.json());
 
-// Serve static assets from the frontend public directory so /icc-logo-no-bg.png and others resolve
+// Serve static assets from the frontend public directory
 app.use(
   express.static(path.join(__dirname, '..', 'public'), {
     maxAge: '7d',
@@ -26,78 +29,11 @@ app.use(
   })
 );
 
+// API Routes
+app.use('/api', meetingsRoutes);
+
 const io = socketIo(server, {
   cors: corsOptions
-});
-
-// In-memory storage for meetings and participants
-const meetings = new Map();
-const participants = new Map();
-
-// Generate a random 6-character meeting code
-function generateMeetingCode() {
-  const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
-  let result = '';
-  for (let i = 0; i < 6; i++) {
-    result += chars.charAt(Math.floor(Math.random() * chars.length));
-  }
-  return result;
-}
-
-// API Routes
-app.get('/health', (req, res) => {
-  res.json({ status: 'ok', timestamp: new Date().toISOString() });
-});
-
-// Create a new meeting
-app.post('/api/meetings', (req, res) => {
-  const { facilitatorName, meetingTitle } = req.body;
-  
-  if (!facilitatorName || !meetingTitle) {
-    return res.status(400).json({ error: 'Facilitator name and meeting title are required' });
-  }
-
-  const meetingCode = generateMeetingCode();
-  const meetingId = uuidv4();
-  
-  const meeting = {
-    id: meetingId,
-    code: meetingCode,
-    title: meetingTitle,
-    facilitator: facilitatorName,
-    participants: [],
-    queue: [],
-    createdAt: new Date().toISOString(),
-    isActive: true
-  };
-  
-  meetings.set(meetingCode, meeting);
-  
-  console.log(`Meeting created: ${meetingCode} - ${meetingTitle} by ${facilitatorName}`);
-  
-  res.json({
-    meetingCode,
-    meetingId,
-    shareUrl: `${req.protocol}://${req.get('host')}/join/${meetingCode}`
-  });
-});
-
-// Get meeting info
-app.get('/api/meetings/:code', (req, res) => {
-  const { code } = req.params;
-  const meeting = meetings.get(code.toUpperCase());
-  
-  if (!meeting) {
-    return res.status(404).json({ error: 'Meeting not found' });
-  }
-  
-  res.json({
-    code: meeting.code,
-    title: meeting.title,
-    facilitator: meeting.facilitator,
-    participantCount: meeting.participants.length,
-    isActive: meeting.isActive
-  });
 });
 
 // Socket.io connection handling
@@ -106,300 +42,73 @@ io.on('connection', (socket) => {
 
   // Join a meeting
   socket.on('join-meeting', (data) => {
-    const { meetingCode, participantName, isFacilitator = false } = data;
-    
-    // Validate input data
-    if (!meetingCode || typeof meetingCode !== 'string') {
-      socket.emit('error', { message: 'Meeting code is required' });
-      return;
+    const result = socketHandlers.handleJoinMeeting(socket, data);
+    if (result) {
+      const { meeting, participant } = result;
+      
+      // Notify all other participants about the new/updated participant
+      socket.to(meeting.code).emit('participant-joined', {
+        participant,
+        participantCount: meeting.participants.length
+      });
+      
+      // Send updated participant list
+      io.to(meeting.code).emit('participants-updated', meeting.participants);
     }
-    
-    if (!participantName || typeof participantName !== 'string' || participantName.trim().length === 0) {
-      socket.emit('error', { message: 'Participant name is required' });
-      return;
-    }
-    
-    // Sanitize participant name
-    const sanitizedName = participantName.trim().substring(0, 50); // Limit length
-    
-    const meeting = meetings.get(meetingCode.toUpperCase());
-    
-    if (!meeting) {
-      socket.emit('error', { message: 'Meeting not found' });
-      return;
-    }
-    
-    // Verify facilitator access - only the original facilitator can join as facilitator
-    if (isFacilitator && sanitizedName !== meeting.facilitator) {
-      console.log(`Unauthorized facilitator attempt: ${sanitizedName} tried to join meeting ${meetingCode} as facilitator (actual facilitator: ${meeting.facilitator})`);
-      socket.emit('error', { message: 'Only the meeting creator can join as facilitator' });
-      return;
-    }
-    
-    // Prevent duplicate joins from the same socket
-    if (participants.has(socket.id)) {
-      const existing = participants.get(socket.id);
-      const existingMeeting = meetings.get(existing.meetingCode);
-      if (existingMeeting) {
-        // Already joined; just re-emit joined event for idempotency
-        socket.emit('meeting-joined', {
-          meeting: {
-            code: existingMeeting.code,
-            title: existingMeeting.title,
-            facilitator: existingMeeting.facilitator
-          },
-          participant: existing.participant,
-          queue: existingMeeting.queue,
-          participants: existingMeeting.participants
-        });
-        return;
-      }
-    }
-
-    // Deduplicate by role/name: if facilitator rejoins, update their socket id instead of adding
-    const existingIndex = meeting.participants.findIndex(p => p.name === sanitizedName && p.isFacilitator === !!isFacilitator);
-    let participant;
-    if (existingIndex !== -1) {
-      participant = {
-        ...meeting.participants[existingIndex],
-        id: socket.id,
-        joinedAt: new Date().toISOString()
-      };
-      meeting.participants[existingIndex] = participant;
-    } else {
-      // Also prevent duplicate participant (non-facilitator) by same name in same meeting
-      const nonFacIndex = meeting.participants.findIndex(p => p.name === sanitizedName && !p.isFacilitator && !isFacilitator);
-      if (nonFacIndex !== -1) {
-        participant = {
-          ...meeting.participants[nonFacIndex],
-          id: socket.id,
-          joinedAt: new Date().toISOString()
-        };
-        meeting.participants[nonFacIndex] = participant;
-      } else {
-        participant = {
-          id: socket.id,
-          name: sanitizedName,
-          isFacilitator,
-          joinedAt: new Date().toISOString(),
-          isInQueue: false,
-          queuePosition: null
-        };
-        meeting.participants.push(participant);
-      }
-    }
-
-    participants.set(socket.id, { meetingCode: meetingCode.toUpperCase(), participant });
-    
-    // Join the meeting room
-    socket.join(meetingCode.toUpperCase());
-    
-    console.log(`${sanitizedName} joined meeting ${meetingCode} as ${isFacilitator ? 'facilitator' : 'participant'}`);
-    
-    // Send meeting info to the participant
-    socket.emit('meeting-joined', {
-      meeting: {
-        code: meeting.code,
-        title: meeting.title,
-        facilitator: meeting.facilitator
-      },
-      participant,
-      queue: meeting.queue,
-      participants: meeting.participants
-    });
-    
-    // Notify all other participants about the new/updated participant
-    socket.to(meetingCode.toUpperCase()).emit('participant-joined', {
-      participant,
-      participantCount: meeting.participants.length
-    });
-    
-    // Send updated participant list
-    io.to(meetingCode.toUpperCase()).emit('participants-updated', meeting.participants);
   });
 
   // Join speaking queue
   socket.on('join-queue', (data) => {
-    const { type = 'speak' } = data; // 'speak', 'direct-response', 'point-of-info', 'clarification'
-    const participantData = participants.get(socket.id);
-    
-    if (!participantData) {
-      socket.emit('error', { message: 'Not in a meeting' });
-      return;
+    const result = socketHandlers.handleJoinQueue(socket, data);
+    if (result) {
+      const { meeting, queueItem } = result;
+      
+      // Notify all participants about queue update
+      io.to(meeting.code).emit('queue-updated', meeting.queue);
+      io.to(meeting.code).emit('participants-updated', meeting.participants);
     }
-    
-    // Validate queue type
-    const validTypes = ['speak', 'direct-response', 'point-of-info', 'clarification'];
-    if (!validTypes.includes(type)) {
-      socket.emit('error', { message: 'Invalid queue type' });
-      return;
-    }
-    
-    const { meetingCode, participant } = participantData;
-    const meeting = meetings.get(meetingCode);
-    
-    if (!meeting) {
-      socket.emit('error', { message: 'Meeting not found' });
-      return;
-    }
-    
-    // Check if already in queue
-    const existingIndex = meeting.queue.findIndex(item => item.participantId === socket.id);
-    if (existingIndex !== -1) {
-      socket.emit('error', { message: 'Already in queue' });
-      return;
-    }
-    
-    const queueItem = {
-      id: uuidv4(),
-      participantId: socket.id,
-      participantName: participant.name,
-      type,
-      timestamp: new Date().toISOString(),
-      position: meeting.queue.length + 1
-    };
-    
-    meeting.queue.push(queueItem);
-    participant.isInQueue = true;
-    participant.queuePosition = meeting.queue.length;
-    
-    console.log(`${participant.name} joined queue for ${type} in meeting ${meetingCode}`);
-    
-    // Notify all participants about queue update
-    io.to(meetingCode).emit('queue-updated', meeting.queue);
-    io.to(meetingCode).emit('participants-updated', meeting.participants);
   });
 
   // Leave speaking queue
   socket.on('leave-queue', () => {
-    const participantData = participants.get(socket.id);
-    
-    if (!participantData) {
-      socket.emit('error', { message: 'Not in a meeting' });
-      return;
-    }
-    
-    const { meetingCode, participant } = participantData;
-    const meeting = meetings.get(meetingCode);
-    
-    if (!meeting) {
-      socket.emit('error', { message: 'Meeting not found' });
-      return;
-    }
-    
-    // Remove from queue
-    const queueIndex = meeting.queue.findIndex(item => item.participantId === socket.id);
-    if (queueIndex !== -1) {
-      meeting.queue.splice(queueIndex, 1);
-      
-      // Update positions for remaining queue items atomically
-      meeting.queue.forEach((item, index) => {
-        item.position = index + 1;
-      });
-      
-      participant.isInQueue = false;
-      participant.queuePosition = null;
-      
-      console.log(`${participant.name} left queue in meeting ${meetingCode}`);
+    const result = socketHandlers.handleLeaveQueue(socket);
+    if (result) {
+      const { meeting } = result;
       
       // Notify all participants about queue update
-      io.to(meetingCode).emit('queue-updated', meeting.queue);
-      io.to(meetingCode).emit('participants-updated', meeting.participants);
+      io.to(meeting.code).emit('queue-updated', meeting.queue);
+      io.to(meeting.code).emit('participants-updated', meeting.participants);
     }
   });
 
   // Facilitator: Next speaker
   socket.on('next-speaker', () => {
-    const participantData = participants.get(socket.id);
-    
-    if (!participantData) {
-      socket.emit('error', { message: 'Not in a meeting' });
-      return;
+    const result = socketHandlers.handleNextSpeaker(socket);
+    if (result) {
+      const { meeting, nextSpeaker } = result;
+      
+      // Notify all participants
+      io.to(meeting.code).emit('next-speaker', nextSpeaker);
+      io.to(meeting.code).emit('queue-updated', meeting.queue);
+      io.to(meeting.code).emit('participants-updated', meeting.participants);
     }
-    
-    const { meetingCode, participant } = participantData;
-    const meeting = meetings.get(meetingCode);
-    
-    if (!meeting || !participant.isFacilitator) {
-      socket.emit('error', { message: 'Not authorized' });
-      return;
-    }
-    
-    if (meeting.queue.length === 0) {
-      socket.emit('error', { message: 'Queue is empty' });
-      return;
-    }
-    
-    // Remove first person from queue atomically
-    const nextSpeaker = meeting.queue.shift();
-    
-    // Update positions for remaining queue items atomically
-    meeting.queue.forEach((item, index) => {
-      item.position = index + 1;
-    });
-    
-    // Update participant status
-    const speakerParticipant = meeting.participants.find(p => p.id === nextSpeaker.participantId);
-    if (speakerParticipant) {
-      speakerParticipant.isInQueue = false;
-      speakerParticipant.queuePosition = null;
-    }
-    
-    console.log(`Next speaker: ${nextSpeaker.participantName} in meeting ${meetingCode}`);
-    
-    // Notify all participants
-    io.to(meetingCode).emit('next-speaker', nextSpeaker);
-    io.to(meetingCode).emit('queue-updated', meeting.queue);
-    io.to(meetingCode).emit('participants-updated', meeting.participants);
   });
 
   // Handle disconnection
   socket.on('disconnect', () => {
-    console.log('User disconnected:', socket.id);
-    
-    const participantData = participants.get(socket.id);
-    if (participantData) {
-      const { meetingCode, participant } = participantData;
-      const meeting = meetings.get(meetingCode);
+    const result = socketHandlers.handleDisconnect(socket);
+    if (result) {
+      const { meeting, participant } = result;
       
-      if (meeting) {
-        // Remove from participants
-        const participantIndex = meeting.participants.findIndex(p => p.id === socket.id);
-        if (participantIndex !== -1) {
-          meeting.participants.splice(participantIndex, 1);
-        }
-        
-        // Remove from queue if present
-        const queueIndex = meeting.queue.findIndex(item => item.participantId === socket.id);
-        if (queueIndex !== -1) {
-          meeting.queue.splice(queueIndex, 1);
-          
-          // Update positions for remaining queue items atomically
-          meeting.queue.forEach((item, index) => {
-            item.position = index + 1;
-          });
-        }
-        
-        console.log(`${participant.name} left meeting ${meetingCode}`);
-        
-        // Notify remaining participants
-        socket.to(meetingCode).emit('participant-left', {
-          participantId: socket.id,
-          participantName: participant.name,
-          participantCount: meeting.participants.length
-        });
-        
-        socket.to(meetingCode).emit('queue-updated', meeting.queue);
-        socket.to(meetingCode).emit('participants-updated', meeting.participants);
-        
-        // Clean up empty meetings
-        if (meeting.participants.length === 0) {
-          meetings.delete(meetingCode);
-          console.log(`Meeting ${meetingCode} deleted (no participants)`);
-        }
-      }
+      // Notify remaining participants
+      socket.to(meeting.code).emit('participant-left', {
+        participantId: socket.id,
+        participantName: participant.name,
+        participantCount: meeting.participants.length
+      });
       
-      participants.delete(socket.id);
+      socket.to(meeting.code).emit('queue-updated', meeting.queue);
+      socket.to(meeting.code).emit('participants-updated', meeting.participants);
     }
   });
 });
@@ -407,7 +116,7 @@ io.on('connection', (socket) => {
 // Serve static files for QR code joining
 app.get('/join/:code', (req, res) => {
   const { code } = req.params;
-  const meeting = meetings.get(code.toUpperCase());
+  const meeting = meetingsService.getMeeting(code);
   
   if (!meeting) {
     return res.status(404).send(`
@@ -435,7 +144,7 @@ const PORT = process.env.PORT || 3000;
 
 server.listen(PORT, '0.0.0.0', () => {
   console.log(`🚀 Stack Facilitation Backend running on port ${PORT}`);
-  console.log(`📊 Health check: http://localhost:${PORT}/health`);
+  console.log(`📊 Health check: http://localhost:${PORT}/api/health`);
   console.log(`🔗 Join meetings: http://localhost:${PORT}/join/MEETINGCODE`);
 });
 
@@ -446,4 +155,3 @@ process.on('SIGTERM', () => {
     console.log('Process terminated');
   });
 });
-
