+++ src/components/MeetingRoom/index.ts
@@ -1,0 +1,7 @@
+export { MeetingRoomRefactored as MeetingRoom } from './MeetingRoomRefactored';
+export { MeetingHeader } from './MeetingHeader';
+export { CurrentSpeakerAlert } from './CurrentSpeakerAlert';
+export { SpeakingQueue } from './SpeakingQueue';
+export { ActionsPanel } from './ActionsPanel';
+export { LoadingState } from './LoadingState';
+export { ErrorState } from './ErrorState';
