+++ src/components/StackKeeper/index.ts
@@ -1,0 +1,7 @@
+export { StackKeeperRefactored as StackKeeper } from './StackKeeperRefactored';
+export { StackHeader } from './StackHeader';
+export { AddParticipant } from './AddParticipant';
+export { CurrentSpeaker } from './CurrentSpeaker';
+export { SpeakingQueue } from './SpeakingQueue';
+export { InterventionsPanel } from './InterventionsPanel';
+export { SpeakingDistribution } from './SpeakingDistribution';
