+++ src/components/ui/sidebar/index.ts
@@ -1,0 +1,42 @@
+export {
+  Sidebar,
+  SidebarContent,
+  SidebarFooter,
+  SidebarGroup,
+  SidebarGroupAction,
+  SidebarGroupContent,
+  SidebarGroupLabel,
+  SidebarHeader,
+  SidebarInput,
+  SidebarInset,
+  SidebarMenu,
+  SidebarMenuAction,
+  SidebarMenuBadge,
+  SidebarMenuButton,
+  SidebarMenuItem,
+  SidebarMenuSkeleton,
+  SidebarMenuSub,
+  SidebarMenuSubButton,
+  SidebarMenuSubItem,
+  SidebarProvider,
+  SidebarRail,
+  SidebarSeparator,
+  SidebarTrigger,
+  useSidebar,
+} from "./context"
+export { Sidebar } from "./sidebar"
+export { SidebarTrigger } from "./sidebar-trigger"
+export { SidebarInset } from "./sidebar-inset"
+export { SidebarContent, SidebarHeader, SidebarFooter } from "./sidebar-content"
+export { SidebarGroup, SidebarGroupLabel, SidebarGroupAction, SidebarGroupContent } from "./sidebar-group"
+export {
+  SidebarMenu,
+  SidebarMenuItem,
+  SidebarMenuButton,
+  SidebarMenuAction,
+  SidebarMenuBadge,
+  SidebarMenuSkeleton,
+} from "./sidebar-menu"
+export { SidebarMenuSub, SidebarMenuSubItem, SidebarMenuSubButton } from "./sidebar-submenu"
+export { SidebarInput, SidebarSeparator } from "./sidebar-input"
+export { SidebarRail } from "./sidebar-rail"
